+++ pages/src/controller/facultySubjectController.php
@@ -9,19 +9,23 @@ class FacultySubjectController extends Controller {
 	}
 
 	public function get_faculty_subjects() {
-
+		$response = $this->model->get_faculty_subjects();
+		return $response;
 	}
 
 	public function add_faculty_subject($data) {
-
+		$response = $this->model->add_faculty_subject($data);
+		return $response;
 	}
 
 	public function update_faculty_subject($data) {
-
+		$response = $this->model->update_faculty_subject($data);
+		return $response;
 	}
 
 	public function delete_faculty_subject($data) {
-
+		$response = $this->model->delete_faculty_subject($data);
+		return $response;
 	}
 
 }